+++ package.json
@@ -18,6 +18,7 @@
     "prepare": "husky install"
   },
   "dependencies": {
+    "@azure/msal-browser": "^2.32.1",
     "@braid/vue-formulate": "^2.5.2",
     "@chatwoot/prosemirror-schema": "https://github.com/chatwoot/prosemirror-schema.git#7e8acadd10d7b932c0dc0bd0a18f804434f83517",
     "@chatwoot/utils": "^0.0.10",
